made spikepanel as separate file
