+++ src/main/java/com/groupesan/project/java/scrumsimulator/mainpackage/ui/panels/SimulationPanel.java
@@ -3,10 +3,7 @@ package com.groupesan.project.java.scrumsimulator.mainpackage.ui.panels;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-
 import javax.swing.*;
-import javax.swing.table.DefaultTableModel;
-import javax.swing.table.TableCellRenderer;
 
 import com.groupesan.project.java.scrumsimulator.mainpackage.impl.BlockerManager;
 import com.groupesan.project.java.scrumsimulator.mainpackage.state.SimulationStateManager;
@@ -20,7 +17,6 @@ public class SimulationPanel extends JPanel implements BaseComponent {
     private JButton stopSimulationButton;
     private JButton showBlockersButton;
     private JButton spikeActivityButton;
-    private DefaultTableModel spikeTableModel;
 
     /** Simulation Panel Initialization. */
     protected SimulationPanel(SimulationStateManager simulationStateManager) {
@@ -41,24 +37,7 @@ public class SimulationPanel extends JPanel implements BaseComponent {
         showBlockersButton.setVisible(false);
         spikeActivityButton.setVisible(false);
 
-     //   GridBagConstraints gbc = new GridBagConstraints();
-//        gbc.insets = new Insets(10, 10, 10, 10);
-//        gbc.fill = GridBagConstraints.HORIZONTAL;
-//
-//        gbc.gridx = 0;
-//        gbc.gridy = 0;
-//        this.add(startSimulationButton, gbc);
-//
-//        gbc.gridx = 1;
-//        this.add(stopSimulationButton, gbc);
-//
-//        gbc.gridx = 0;
-//        gbc.gridy = 1;
-//        this.add(showBlockersButton, gbc);
-//
-//        gbc.gridx = 1;
         spikeActivityButton.setBackground(Color.YELLOW);
-//        this.add(spikeActivityButton, gbc);
 
         startSimulationButton.addActionListener(new ActionListener() {
             @Override
@@ -100,13 +79,15 @@ public class SimulationPanel extends JPanel implements BaseComponent {
         spikeActivityButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                showSpikeLogPanel();
+                SpikePanel spikePanel = new SpikePanel();
+                spikePanel.showSpikeLogFrame();
             }
         });
-    add(startSimulationButton);
-    add(stopSimulationButton);
-    add(showBlockersButton);
-    add(spikeActivityButton);
+
+        add(startSimulationButton);
+        add(stopSimulationButton);
+        add(showBlockersButton);
+        add(spikeActivityButton);
     }
 
     private void updateButtonVisibility() {
@@ -125,54 +106,5 @@ public class SimulationPanel extends JPanel implements BaseComponent {
             showBlockersButton.setBackground(Color.RED);
             showBlockersButton.setText("Resolve Blockers");
         }
-    }
-
-    private void showSpikeLogPanel() {
-        JPanel spikeLogPanel = new JPanel(new BorderLayout());
-
-        String[] columns = {"S.No", "Spike Title", "Reason/Desc", "Resolved"};
-
-        spikeTableModel = new DefaultTableModel(columns, 0) {
-            @Override
-            public Class<?> getColumnClass(int column) {
-                return column == 3 ? Boolean.class : String.class;
-            }
-        };
-
-        JTable spikeTable = new JTable(spikeTableModel);
-
-        spikeTable.getColumnModel().getColumn(3).setCellEditor(new DefaultCellEditor(new JCheckBox()));
-        spikeTable.getColumnModel().getColumn(3).setCellRenderer(new TableCellRenderer() {
-            @Override
-            public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected, boolean hasFocus, int row, int column) {
-                JCheckBox checkBox = new JCheckBox();
-                checkBox.setSelected(value != null && (Boolean) value);
-                return checkBox;
-            }
-        });
-
-        spikeTable.getModel().addTableModelListener(e -> {
-            int row = e.getFirstRow();
-            int col = e.getColumn();
-            if (col == 3 && (Boolean) spikeTableModel.getValueAt(row, col)) {
-                spikeTableModel.removeRow(row);
-            }
-        });
-
-        JScrollPane scrollPane = new JScrollPane(spikeTable);
-
-        JButton addNewSpikeButton = new JButton("Add Spike");
-        addNewSpikeButton.addActionListener(e -> {
-            // Implement logic to add a new spike
-        });
-
-        spikeLogPanel.add(scrollPane, BorderLayout.CENTER);
-        spikeLogPanel.add(addNewSpikeButton, BorderLayout.SOUTH);
-
-        JFrame spikeLogFrame = new JFrame("Spike Activity Log");
-        spikeLogFrame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-        spikeLogFrame.setSize(600, 400);
-        spikeLogFrame.add(spikeLogPanel);
-        spikeLogFrame.setVisible(true);
     }
 }
